chore : env 관련 설정 추가
