+++ src/utils/uploadFileToStorage.tsx
@@ -1,0 +1,3 @@
+export async function uploadFileToStorage() {
+	const supabase = await createServerS
+}
