+++ tsconfig.app.json
@@ -28,5 +28,5 @@
 			"@/*": ["src/*"] // '@'를 'src' 폴더로 매핑
 		}
 	},
-	"include": ["src"]
+	"include": ["vite-env.d.ts", "src"]
 }
