+++ tsconfig.json
@@ -6,6 +6,7 @@
 		"paths": {
 			"@/*": ["src/*"] // '@'를 'src' 폴더로 매핑
 		},
-		"types": ["vite-plugin-svgr/client"]
+		"types": ["vite-plugin-svgr/client"],
+		"esModuleInterop": true
 	}
 }
